DB transfor test 1

+++ db/migrate/20131124000748_add_columns_to_loggings.rb
@@ -1,6 +1,6 @@
 class AddColumnsToLoggings < ActiveRecord::Migration
   def change
     add_column :loggings, :box_status, :integer
-    add_column :loggings, :action_type, :tinyint
+    add_column :loggings, :action_type, :integer
   end
 end
